修改安全规范告警问题

+++ frameworks/core/components_ng/pattern/rich_editor/rich_editor_pattern.h
@@ -137,7 +137,7 @@ public:
         uint32_t spanTextLength = 0;
         for (auto& span : spans_) {
             spanTextLength += StringUtils::ToWstring(span->content).length();
-            span->position = spanTextLength;
+            span->position = static_cast<int32_t>(spanTextLength);
         }
     }
